Match the Fcl combine state machine 002f6cf0
The parked 0-word draft now has one signature per callee:

- 00313ae0 is (s8 kind, u16 id) with a widened copy of the id; 00313a80
  and 003139d0 are (s8, s8).
- 003147d0 returns u8 *.
- 0011d140 is (u8 *, s32 color) and forwards the colour to the
  now-prototyped 0011e230. 0011d170 does the same with arg1.
- 002bb1c0 returns s32 from an s8 index.
- 003144d0 takes the slot byte as s8.
- The unit uses the u8 view of the colour packers, and the alphas in
  002f6cf0 and 002f9d90 are u8 locals.

Every affected unit keeps its statuses. y_fclCombine goes from 35 to 36
MATCH.

+++ src/Event/Fcl/y_fclItemShopDraw.c
@@ -130,7 +130,7 @@ extern f32 iGpffff84cc;
 extern f32 D_0064A210[];
 extern f32 D_0064A218[];
 extern void func_0045af60(s32, s32, s32, s32);
-extern s8 func_002bb1c0(s8 arg0);
+extern s32 func_002bb1c0(s8 arg0);
 extern void func_00106620(s16 arg0, s32 arg1);
 extern void func_002e7a80(s32 arg0);
 void func_0033c490(u8 *arg0);
